Practicing a little bit more.
Adding a member to add an item without resizing the array.
Making a helper function not to repeat myself when adding an item
both at the beginning as at the end of the Array.
And the constructor which initializes an Array from an initializer list
makes confusion when making an Array with a fixed size and with an
initializer list, so for now it's commented, but must be fixed.