More practice. See below:
I've learned the copy-and-swap idiom, and the const-and-back-again one,
so in this modification they both were put in practice.
Other things such as adding the copy constructor were added.
I wasn't capable of implementing it before because I had a constructor
which received a std::initializer_list and that made confusion when
making an Array object, i.e, the compiler *always* called the
constructor which made an Array with a set of items even if you just
passed it an item pretending to make an Array with a fixed size such as
5. I'd like to have both constructor, so if I want to make an Array with
a fixed size, I could. Also if I want to make an Array with a set of
items, I could too.