Fixed constructor with initializer list.
To solve the confusion, if you, indeed, want to make an Array of N
items, Array(size_t size), instead of an Array with just one item with value N,
Array(std::initializer_list<T>), use parenthesis, because the compiler *will*
call the constructor for an initializer list when using curly braces,
if there's one. If not, it will call an adequate constructor, depending on the
type of the argument(s).